+++ packages/angular-workspace/ok-angular/fv-accordion-item/public-api.ts
@@ -1,0 +1,2 @@
+export * from './fv-accordion-item.directive';
+export * from './fv-accordion-item.module';