+++ packages/ok-components/src/all-components.ts
@@ -6,5 +6,6 @@
 
 import '@ni/spright-components/dist/esm/all-components';
 
+import './fv-accordion-item';
 import './button';
 import './icon-dynamic';