+++ packages/react-workspace/ok-react/src/fv-accordion-item/index.ts
@@ -1,0 +1,7 @@
+import { FvAccordionItem, fvAccordionItemTag } from '@ni/ok-components/dist/esm/fv-accordion-item';
+import { FvAccordionItemAppearance } from '@ni/ok-components/dist/esm/fv-accordion-item/types';
+import { wrap } from '../utilities/react-wrapper';
+
+export { fvAccordionItemTag };
+export { type FvAccordionItem, FvAccordionItemAppearance };
+export const OkFvAccordionItem = wrap(FvAccordionItem);