+++ packages/react-workspace/react-client-app/src/components/App.tsx
@@ -40,6 +40,7 @@ import { TooltipSection } from './TooltipSection';
 import { TreeViewSection } from './TreeViewSection';
 import { ChatConversationSection } from './ChatConversationSection';
 import { RectangleSection } from './RectangleSection';
+import { OkAccordionItemSection } from './OkAccordionItemSection';
 import { OkButtonSection } from './OkButtonSection';
 
 export function App(): React.JSX.Element {
@@ -107,6 +108,7 @@ export function App(): React.JSX.Element {
                                 <TreeViewSection />
                                 <ChatConversationSection />
                                 <RectangleSection />
+                                <OkAccordionItemSection />
                                 <OkButtonSection />
                             </div>
                         </div>